move secagg related methods to a secagg client wrapper
