+++ src/py/flwr/client/__init__.py
@@ -29,4 +29,5 @@ __all__ = [
     "Client",
     "KerasClient",
     "NumPyClient",
+    "SecAggClient",
 ]
